Add RESTful endpoints

+++ src/main/java/resu/muffin/users_and_comments/users/db/RoleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import resu.muffin.users_and_comments.users.entities.Role;
 
 @Repository
-public interface RoleRepository extends CrudRepository<Role, Long> {
+public interface RoleRepository extends CrudRepository<Role, Short> {
     Role findByName(String name);
 }
